requirements.txt
aiohttp==3.11.11
alembic==1.14.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.37

